requirements.txt
# local package
-e .

# external requirements
certifi==2021.10.8
numpy==1.22.3
pandas==1.4.2
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
wincertstore==0.2
# bunch==1.0.1
tensorflow==2.8.0
matplotlib==3.5.1
pillow==9.1.0
opencv-python==4.5.5.64
tqdm==4.64.0
scikit-image==0.19.2
munch==2.5.0
scipy==1.8.0